+++ src/main/java/net/kaikk/mc/bcl/Exceptions/NegativeValueException.java
@@ -1,0 +1,13 @@
+package net.kaikk.mc.bcl.Exceptions;
+
+public class NegativeValueException extends MyException {
+
+    public NegativeValueException()
+    {
+    }
+
+    public NegativeValueException(String message)
+    {
+        super(message);
+    }
+}